+++ lib/active_model/serializer.rb
@@ -1,95 +1,7 @@
 require "active_support/core_ext/class/attribute"
 require "active_support/core_ext/module/anonymous"
-require "set"
 
 module ActiveModel
-  class OrderedSet
-    def initialize(array)
-      @array = array
-      @hash = {}
-
-      array.each do |item|
-        @hash[item] = true
-      end
-    end
-
-    def merge!(other)
-      other.each do |item|
-        next if @hash.key?(item)
-
-        @hash[item] = true
-        @array.push item
-      end
-    end
-
-    def to_a
-      @array
-    end
-  end
-
-  # Active Model Array Serializer
-  #
-  # It serializes an Array, checking if each element that implements
-  # the +active_model_serializer+ method.
-  #
-  # To disable serialization of root elements:
-  #
-  #     ActiveModel::ArraySerializer.root = false
-  #
-  class ArraySerializer
-    attr_reader :object, :options
-
-    class_attribute :root
-
-    def initialize(object, options={})
-      @object, @options = object, options
-    end
-
-    def serializable_array
-      @object.map do |item|
-        if @options.has_key? :each_serializer
-          serializer = @options[:each_serializer]
-        elsif item.respond_to?(:active_model_serializer)
-          serializer = item.active_model_serializer
-        end
-
-        if serializer
-          serializer.new(item, @options)
-        else
-          item
-        end
-      end
-    end
-
-    def as_json(options = nil)
-      @options[:hash] = hash = {}
-      @options[:unique_values] = {}
-
-      options ||= {}
-      # by default, don't include root values for objects inside the array
-      options[:root] = false if options[:root].nil?
-
-      array = serializable_array.map do |item|
-        if item.is_a?(::ActiveModel::Serializer) ||
-            item.is_a?(::ActiveModel::ArraySerializer)
-          # serialize AMS using as_json to allow options
-          item.as_json(options)
-        elsif item.respond_to?(:serializable_hash)
-          # otherwise, prefer serializable_hash over as_json if available
-          item.serializable_hash
-        else
-          item.as_json(options)
-        end
-      end
-
-      if root = @options[:root]
-        hash.merge!(root => array)
-      else
-        array
-      end
-    end
-  end
-
   # Active Model Serializer
   #
   # Provides a basic serializer implementation that allows you to easily
@@ -124,6 +36,9 @@ module ActiveModel
   #     end
   #
   class Serializer
+    INCLUDE_METHODS = {}
+    INSTRUMENT = { :serialize => :"serialize.serializer", :associations => :"associations.serializer" }
+
     class IncludeError < StandardError
       attr_reader :source, :association
 
@@ -133,178 +48,6 @@ module ActiveModel
 
       def to_s
         "Cannot serialize #{association} when #{source} does not have a root!"
-      end
-    end
-
-    module Associations #:nodoc:
-      class Config #:nodoc:
-        class_attribute :options
-
-        def self.refine(name, class_options)
-          current_class = self
-
-          Class.new(self) do
-            singleton_class.class_eval do
-              define_method(:to_s) do
-                "(subclass of #{current_class.name})"
-              end
-
-              alias inspect to_s
-            end
-
-            self.options = class_options
-          end
-        end
-
-        self.options = {}
-
-        def initialize(name, source, options={})
-          @name = name
-          @source = source
-          @options = options
-        end
-
-        def option(key, default=nil)
-          if @options.key?(key)
-            @options[key]
-          elsif self.class.options.key?(key)
-            self.class.options[key]
-          else
-            default
-          end
-        end
-
-        def target_serializer
-          option(:serializer)
-        end
-
-        def source_serializer
-          @source
-        end
-
-        def key
-          option(:key) || @name
-        end
-
-        def root
-          option(:root) || plural_key
-        end
-
-        def name
-          option(:name) || @name
-        end
-
-        def associated_object
-          option(:value) || source_serializer.send(name)
-        end
-
-        def embed_ids?
-          option(:embed, source_serializer._embed) == :ids
-        end
-
-        def embed_objects?
-          option(:embed, source_serializer._embed) == :objects
-        end
-
-        def embed_in_root?
-          option(:include, source_serializer._root_embed)
-        end
-
-        def embeddable?
-          !associated_object.nil?
-        end
-
-      protected
-
-        def find_serializable(object)
-          if target_serializer
-            target_serializer.new(object, source_serializer.options)
-          elsif object.respond_to?(:active_model_serializer) && (ams = object.active_model_serializer)
-            ams.new(object, source_serializer.options)
-          else
-            object
-          end
-        end
-      end
-
-      class HasMany < Config #:nodoc:
-        alias plural_key key
-
-        def serialize
-          associated_object.map do |item|
-            find_serializable(item).serializable_hash
-          end
-        end
-        alias serialize_many serialize
-
-        def serialize_ids
-          # Use pluck or select_columns if available
-          # return collection.ids if collection.respond_to?(:ids)
-
-          associated_object.map do |item|
-            item.read_attribute_for_serialization(:id)
-          end
-        end
-      end
-
-      class HasOne < Config #:nodoc:
-        def embeddable?
-          if polymorphic? && associated_object.nil?
-            false
-          else
-            true
-          end
-        end
-
-        def polymorphic?
-          option :polymorphic
-        end
-
-        def polymorphic_key
-          associated_object.class.to_s.demodulize.underscore.to_sym
-        end
-
-        def plural_key
-          if polymorphic?
-            associated_object.class.to_s.pluralize.demodulize.underscore.to_sym
-          else
-            key.to_s.pluralize.to_sym
-          end
-        end
-
-        def serialize
-          object = associated_object
-
-          if object && polymorphic?
-            {
-              :type => polymorphic_key,
-              polymorphic_key => find_serializable(object).serializable_hash
-            }
-          elsif object
-            find_serializable(object).serializable_hash
-          end
-        end
-
-        def serialize_many
-          object = associated_object
-          value = object && find_serializable(object).serializable_hash
-          value ? [value] : []
-        end
-
-        def serialize_ids
-          object = associated_object
-
-          if object && polymorphic?
-            {
-              :type => polymorphic_key,
-              :id => object.read_attribute_for_serialization(:id)
-            }
-          elsif object
-            object.read_attribute_for_serialization(:id)
-          else
-            nil
-          end
-        end
       end
     end
 
@@ -360,6 +103,9 @@ module ActiveModel
 
       def define_include_method(name)
         method = "include_#{name}?".to_sym
+
+        INCLUDE_METHODS[name] = method
+
         unless method_defined?(method)
           define_method method do
             true
@@ -423,16 +169,31 @@ module ActiveModel
         klass = model_class
         columns = klass.columns_hash
 
-        attrs = _attributes.inject({}) do |hash, (name,key)|
-          column = columns[name.to_s]
-          hash.merge key => column.type
+        attrs = {}
+        _attributes.each do |name, key|
+          if column = columns[name.to_s]
+            attrs[key] = column.type
+          else
+            # Computed attribute (method on serializer or model). We cannot
+            # infer the type, so we put nil.
+            attrs[key] = nil
+          end
         end
 
-        associations = _associations.inject({}) do |hash, (attr,association_class)|
+        associations = {}
+        _associations.each do |attr, association_class|
           association = association_class.new(attr, self)
 
-          model_association = klass.reflect_on_association(association.name)
-          hash.merge association.key => { model_association.macro => model_association.name }
+          if model_association = klass.reflect_on_association(association.name)
+            # Real association.
+            associations[association.key] = { model_association.macro => model_association.name }
+          else
+            # Computed association. We could infer has_many vs. has_one from
+            # the association class, but that would make it different from
+            # real associations, which read has_one vs. belongs_to from the
+            # model.
+            associations[association.key] = nil
+          end
         end
 
         { :attributes => attrs, :associations => associations }
@@ -458,6 +219,7 @@ module ActiveModel
       def root(name)
         self._root = name
       end
+      alias_method :root=, :root
 
       def inherited(klass) #:nodoc:
         return if klass.anonymous?
@@ -513,7 +275,7 @@ module ActiveModel
     end
 
     def include?(name)
-      send "include_#{name}?".to_sym
+      send INCLUDE_METHODS[name]
     end
 
     def include!(name, options={})
@@ -563,7 +325,7 @@ module ActiveModel
         if association.embed_in_root? && hash.nil?
           raise IncludeError.new(self.class, association.name)
         elsif association.embed_in_root? && association.embeddable?
-          merge_association hash, association.root, association.serialize_many, unique_values
+          merge_association hash, association.root, association.serializables, unique_values
         end
       elsif association.embed_objects?
         node[association.key] = association.serialize
@@ -579,13 +341,15 @@ module ActiveModel
     # a unique list of all of the objects that are already in the Array. This
     # avoids the need to scan through the Array looking for entries every time
     # we want to merge a new list of values.
-    def merge_association(hash, key, value, unique_values)
-      if current_value = unique_values[key]
-        current_value.merge! value
-        hash[key] = current_value.to_a
-      elsif value
-        hash[key] = value
-        unique_values[key] = OrderedSet.new(value)
+    def merge_association(hash, key, serializables, unique_values)
+      already_serialized = (unique_values[key] ||= {})
+      serializable_hashes = (hash[key] ||= [])
+
+      serializables.each do |serializable|
+        unless already_serialized.include? serializable.object
+          already_serialized[serializable.object] = true
+          serializable_hashes << serializable.serializable_hash
+        end
       end
     end
 
@@ -611,7 +375,8 @@ module ActiveModel
     # Use ActiveSupport::Notifications to send events to external systems.
     # The event name is: name.class_name.serializer
     def instrument(name, payload = {}, &block)
-      ActiveSupport::Notifications.instrument("#{name}.serializer", payload, &block)
+      event_name = INSTRUMENT[name]
+      ActiveSupport::Notifications.instrument(event_name, payload, &block)
     end
   end
 end